Parameterize objects, classes and members by the signature (#4)
## Parameterize object, classes and members by signature
When writing the counter application, I've realized that having all
types existentially quantified was not practical for writing a class.
I've introduced the concept of Signature. A signature contains the types
of the public fields, private fields and the class label.

## Universal Counter example
I've implemented the counter class, which has a single public field of
type Nat. It supports initialization and increment by a certain value.