Set relaxedAutoImplicit to false (#7)
I think this is a saner default, and I have already caught one typo with
it.

From the documentation:

> Automatic implicit parameters insertion is controlled by two options.
By default, automatic implicit parameter insertion is relaxed, which
means that any unbound identifier may be a candidate for automatic
insertion. Setting the option
[relaxedAutoImplicit](https://lean-lang.org/doc/reference/latest/Definitions/Headers-and-Signatures/#relaxedAutoImplicit)
to
[false](https://lean-lang.org/doc/reference/latest/Basic-Types/Booleans/#Bool___false)
disables relaxed mode and causes only identifiers that consist of a
single character followed by zero or more digits to be considered for
automatic insertion.