Add dynamic label (#27)
Adds the possibility to add dynamic data to the label.

Currently when translating from an Object to a Resource we put the
object label as the resource label. However, in some cases the label we
want to put in the Resource depends on the contents of the object. This
pr extends the label definition to allow that