Intents (#3)
- Adds intents to the GOOSE prototype.
- Intents are compiled to intent resources. Two RLs are generated: one
for the intent logic checked on intent resource consumption, and one
intent member logic for the objects consumed by the intent (checking
that the right intent is created). The class of any object consumed by
the intent needs to know about the intent and include the appropriate
intent id among its member ids, so that the intent member logic can be
selected when consuming the object of the class to create the intent
resource.
- The intents in this PR use fake nullifiers (to be addressed in a
future PR).

Example intent from the Kudos app:
```
  intent SwapIntent (wantOriginator : ExternalIdentity) (wantQuantity : Nat)
   intentLogic 
      in (give : Kudos)  -- List of resources that we consume
      want (createdKudos : Set Kudo) (unused : Set OtherObject) : Bool := 
          let kudosToGet : Nat := 
               createdKudos
              |> select \{x => owner x == owner give && originator x == wantOriginator }
              |> map quantity
              |> sum
          in kudosToGet >= wantQuantity;
```