Destructors (#30)
This pr adds destructurs as new members of a class.
On the user level, a destructor definition consists only of the
`invariant` specification, which takes self as an object and additional
custom arguments.
I've added an example destructor in the owned counter example. The
destructor only allows the counter to be destroyed if it has reached a
value of at least 10.

---------

Co-authored-by: Łukasz Czajka <62751+lukaszcz@users.noreply.github.com>