Allow functions to destroy objects (#62)
This pr extends ecosystem functions with the ability to destroy objects.
A destroyed object will be added to the list of consumed resources in
the action and an ephemeral counterpart will be added to the list of
created resources in order for the balance check to pass.

The universal counter example has been extended to showcase the new
feature.
- *mutual* increment. A new function that increments each counter with
the value of the other.
- *absorb*. A new function that increments a counter by the value of the
other, which is destroyed.

---------

Co-authored-by: Łukasz Czajka <62751+lukaszcz@users.noreply.github.com>