Allow functions to construct new objects and specify which selves are destroyed/consumed (#65)
# Changes in the model:
1. Functions can construct new objects
2. Functions specify for each object argument whether it should be
destroyed or disassembled. Destroyed arguments are balanced with
ephemeral resources automatically. Disassembled arguments are meant to
be manually balanced by the user in the function.

# Apps:
I've extended the kudos bank example with a new class `Check`. I've
added two functions to the ecosystem:
1. `IssueCheck`. It subtracts some amount from your bank account and
creates a new check for that amount.
2. `DepositCheck`. It destroys a check and adds its amount to your bank
account.