Auction example (#74)
This pr extends the KudosBank example to allow simple Auctions.

Any user can start an auction by providing a check (to be auctioned) and
specifying which denomination they want in return.

Then any other user is allowed to place bids. As expected, a bid is only
accepted if it is higher than the current highest bid. Every time a bid
is accepted, the previous highest bidder receives their kudos back in
the form of a check.

The owner can decide to close the auction. Then the original auctioned
check is transferred to the highest bidder. The owner gets the check
corresponding to the highest bid.