Nested calls (#81)
- Adds `Class.Member.Body` data structure which represents the body of a
class member (constructor, destructor, method). The body is a sequence
of statements:
  1. constructor call,
  2. destructor call,
  3. method call,
  4. fetch object by id,
  5. return value.
- Object changes within a single body are not tracked for fetches in
this body, i.e., every fetch retrieves the value of the object at the
entrance to the method ignoring subsequent changes performed by method
calls in the body.
- Updates Tasks and the translation to enable translating members with
complex bodies. In particular, this enables translating nested calls.
- Updates Applib and Apps accordingly.
- Adds a TwoCounter app based on
[TwoCounter.lean](https://github.com/anoma/goose-lean/blob/7581ca5e79ca6b56836ed6a83db872fb65a1220d/Apps/TwoCounter.lean)
by @janmasrovira. A `TwoCounter` object has references (`ObjectId`) to
two `Counter` objects. There is a mutual increment (`incrementBoth`)
method which fetches the values of counters and calls Counter.increment
on the counter ids with increment value dependent on counter values.

---------

Co-authored-by: Jan Mas Rovira <janmasrovira@gmail.com>