feat: Scopes (#118)
Allow programs to call methods from different ecosystems.

I've implemented a custom tactic so that the changes are transparent to
the surface language for constructors, methods, destructors and
upgrades.

There are some things that need to be improved:
1. Only works if the `EcosystemIdEnum : FinEnum EcosystemId` is given
explicitly and not inferred as an instance. We need to figure out how to
properly reduce an expression with instance arguments in it.
2. Multimethods require the user to provide the ecosystem id explicitly
in the code. In the future this should be automatized too.

---------

Co-authored-by: Lukasz Czajka <lukasz@heliax.dev>