Debug (#132)
This PR fixes several bugs - mostly involving mismatches in “serialized”
types. To make debugging easier, it changes the signature of the
resource-logic function: it is now monadic, maintains a stack trace, and
returns a descriptive error instead of simply `false` on failure.

I also fixed the monadic version of the `check` macro, which was not
working correctly. I had to rename it to `docheck` to avoid conflicts
with the term-level `check`.

Finally, I added a rather hacky mechanism, `mydefaultM`, to report the
location of logic-function errors even when there is no explicit
`failwith` clause. We should keep this for now but plan to remove it
later.

I've also added a tiny example - a program with a single constructor
call - that runs on CI.